feature: custom sec-websocket-key in client.
This gives the caller the ability to explicitly set the value of the
Sec-WebSocket-Key header instead of having it generated randomly. For
the most part this is not needed, but it is helpful for certain
use-cases like integration tests and reverse proxy applications that
need to be as transparent as possible with proxying websocket
connections.